Extract Search class
